# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_AVGsync_keyfreestay_backend_internal_model_request.CreateBookingRequest:
    properties:
      check_in:
        description: Дата заезда в формате YYYY-MM-DD.
        example: "2026-06-10"
        type: string
      check_out:
        description: Дата выезда в формате YYYY-MM-DD. Должна быть позже даты заезда.
        example: "2026-06-15"
        type: string
      guests_count:
        description: Количество гостей. Не должно превышать вместимость объекта.
        example: 2
        type: integer
      housing_id:
        description: UUID опубликованного объекта жилья.
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      payment_method_id:
        description: UUID платёжного метода пользователя. Поле необязательно.
        example: 660e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - check_in
    - check_out
    - guests_count
    - housing_id
    type: object
...
  github_com_AVGsync_keyfreestay_backend_internal_model_response.BookingListItem:
    properties:
      check_in:
        description: Дата заезда.
        example: "2026-06-10T00:00:00Z"
        type: string
      check_out:
        description: Дата выезда.
        example: "2026-06-15T00:00:00Z"
        type: string
      created_at:
        description: Дата создания бронирования.
        example: "2026-05-15T12:00:00Z"
        type: string
      housing_id:
        description: UUID объекта жилья.
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      id:
        description: UUID бронирования.
        example: 770e8400-e29b-41d4-a716-446655440000
        type: string
      status:
        description: Статус бронирования.
        enum:
        - pending
        - confirmed
        - cancelled
        - completed
        example: pending
        type: string
      total_price:
        description: Итоговая стоимость бронирования.
        example: 33000
        type: number
    type: object
  github_com_AVGsync_keyfreestay_backend_internal_model_response.BookingListResponse:
    properties:
      items:
        description: Бронирования текущего пользователя.
        items:
          $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_response.BookingListItem'
        type: array
      total:
        description: Общее количество элементов в ответе.
        example: 3
        type: integer
    type: object
  github_com_AVGsync_keyfreestay_backend_internal_model_response.BookingResponse:
    properties:
      check_in:
        description: Дата заезда.
        example: "2026-06-10T00:00:00Z"
        type: string
      check_out:
        description: Дата выезда.
        example: "2026-06-15T00:00:00Z"
        type: string
      created_at:
        description: Дата создания бронирования.
        example: "2026-05-15T12:00:00Z"
        type: string
      guests_count:
        description: Количество гостей.
        example: 2
        type: integer
      housing_id:
        description: UUID объекта жилья.
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      id:
        description: UUID бронирования.
        example: 770e8400-e29b-41d4-a716-446655440000
        type: string
      payment_method_id:
        description: UUID платёжного метода, если он был выбран.
        example: 660e8400-e29b-41d4-a716-446655440000
        type: string
      price_per_night_snapshot:
        description: Цена за ночь, зафиксированная на момент бронирования.
        example: 6500
        type: number
      service_fee:
        description: Сервисный сбор.
        example: 500
        type: number
      status:
        description: Статус бронирования.
        enum:
        - pending
        - confirmed
        - cancelled
        - completed
        example: pending
        type: string
      total_price:
        description: Итоговая стоимость бронирования.
        example: 33000
        type: number
    type: object
...
info:
  contact:
    email: support@keyfreestay.local
    name: KeyFreeStay API Support
  description: |-
    REST API сервиса KeyFreeStay для краткосрочной аренды жилья без ключей.

    API покрывает регистрацию и вход, профиль пользователя, объекты жилья, изображения жилья, бронирования, платёжные методы и публичную контактную заявку.
    Схема авторизации: вызовите `POST /auth/login`; API установит JWT в HttpOnly cookie `token`. Защищённые маршруты читают JWT из cookie.
    Swagger UI доступен только при `DEBUG=true`: `/swagger`, JSON спецификация — `/swagger/doc.json`.
  title: KeyFreeStay API
  version: "1.0"
paths:
...
  /booking:
    get:
      description: Возвращает все бронирования текущего пользователя, отсортированные
        по дате заезда от новых к старым.
      produces:
      - application/json
      responses:
        "200":
          description: Список бронирований пользователя
          schema:
            $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_response.BookingListResponse'
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to get booking list — список не получен
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Получить список бронирований
      tags:
      - Бронирования
  /booking/{id}:
    delete:
      description: |-
        Удаляет бронирование текущего пользователя по ID.
        Пользователь может удалить только своё бронирование.
      parameters:
      - description: UUID бронирования
        example: 550e8400-e29b-41d4-a716-446655440000
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Пустой ответ со статусом 200
          schema:
            type: string
        "400":
          description: missing booking ID — id не передан в path
          schema:
            type: string
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to delete booking — бронирование не найдено, не принадлежит
            пользователю или не удалено
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Удалить бронирование
      tags:
      - Бронирования
    get:
      description: |-
        Возвращает бронирование текущего пользователя.
        Пользователь может получить только своё бронирование.
      parameters:
      - description: UUID бронирования
        example: 550e8400-e29b-41d4-a716-446655440000
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Детали бронирования
          schema:
            $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_response.BookingResponse'
        "400":
          description: missing booking ID — id не передан в path
          schema:
            type: string
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to get booking — бронирование не найдено или не принадлежит
            пользователю
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Получить бронирование по ID
      tags:
      - Бронирования
  /booking/create:
    put:
      consumes:
      - application/json
      description: |-
        Создаёт бронирование опубликованного объекта жилья для текущего пользователя.
        Объект должен быть опубликован, вместимость должна быть не меньше `guests_count`, даты не должны пересекаться с активными бронированиями.
        Стоимость рассчитывается сервером: цена за ночь на момент бронирования * количество ночей + сервисный сбор.
      parameters:
      - description: Данные бронирования
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_request.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Созданное бронирование
          schema:
            $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_response.BookingResponse'
        "400":
          description: invalid request body — тело запроса не JSON или структура не
            совпадает со схемой
          schema:
            type: string
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to create booking — объект не найден, не опубликован,
            занят, не подходит по гостям или БД отклонила запрос
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Создать бронирование
      tags:
      - Бронирования
...
schemes:
- http
securityDefinitions:
  CookieAuth:
    description: 'Cookie авторизации. Формат заголовка: Cookie: token={jwt}'
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"